软件正式版本1.3.0 双功能局放测试仪的一个成熟版本，时间：2017.6.2
1.具备基本的TEV和超声局放测试功能（db值显示，峰值保持，颜色报警，TEV严重度，超声耳机输出）
2.TEV模式下能选择显示PRPS图，PRPD图和柱状图，超声模式下能显示PRPS图
3.具有手动录波和故障自动录波功能，能够自动保存数据，能够在设备上图形化查看数据，超声数据能够使用声音播放器回放
4.具有空间自我管理功能，超过设定值会删除陈旧波形文件

+++ Gui/MainMenu/Menu6/DebugSet/recwaveform.h
@@ -38,7 +38,7 @@ private:
 
     MODE mode;
 
-    int x, min, max;
+    double x, min, max;
 
     double scale;
